+++ test/runtests.jl
@@ -10,12 +10,6 @@ using SparseArrays
 
 import MatrixDepot: DataError
 
-function toggle_db()
-    MatrixDepot.toggle_remote()
-    MatrixDepot.init(ignoredb=true)
-    MatrixDepot.update()
-end
-
 # that will download the index files if necessary and initialize internal data
 MatrixDepot.init()
 
@@ -38,18 +32,9 @@ include("include_generator.jl")
         include(tp)
         println("finished $(tp)")
     end
-
-    xdir = MatrixDepot.data_dir()
-    xtmpdir = string(xdir, ".tmp")
-    println("mv $xdir $xtmpdir")
-    mv(xdir, xtmpdir, force=true)
-    toggle_db()
-    @test mdopen(sp(1)) !== nothing
-    mv(xtmpdir, xdir, force=true)
 end
 
 @testset "MatrixDepot real remote matrix tests" begin
-    toggle_db()
     println("running remote.jl ...")
     include("remote.jl")
     println("finished remote.jl")